Adds a entry point for load-server-identities error
